requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
async-timeout==4.0.1
attrs==21.2.0
black==21.11b1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configparser==5.2.0
cycler==0.11.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
flake8==4.0.1
fonttools==4.28.3
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
hydra-colorlog==1.1.0
hydra-core==1.1.1
idna==3.3
imageio==2.13.1
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.0
mccabe==0.6.1
multidict==5.2.0
munch==2.5.0
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.4
pathspec==0.9.0
pathtools==0.1.2
Pillow==8.4.0
platformdirs==2.4.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyDeprecate==0.3.1
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytorch-lightning==1.5.4
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.15.2
rsa==4.8
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
segmentation-models-pytorch==0.2.1
sentry-sdk==1.5.0
setuptools-scm==6.3.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
tomli==1.2.2
torch==1.10.0
torchmetrics==0.6.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
wandb==0.12.7
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
